Insertar objetos DONE

+++ Practica #10/src/main/java/practica10/entidades/FamiliaEquipo.java
@@ -13,7 +13,7 @@ public class FamiliaEquipo implements Serializable {
     @Id
     @GeneratedValue
     private
-    long id;
+    int id;
     private String nombre;
 
     public String getNombre() {
@@ -28,7 +28,7 @@ public class FamiliaEquipo implements Serializable {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
